fix jest unittests
